Move delete project to the plugin
